Refactor module structure
- Split __init__.py into XInclude and DocBook specific parts
